Refactor includes
try to always use quotes for project local and angle brackets for
libraries

Aim for this order:
1 SYCL
2 std library
3 other libaries (including portfft for tests)
4 project local
